fix query params dyn routes
