+++ front/pages/api/v1/w/[wId]/data_sources/[name]/databases/[dId]/tables/[tId].ts
@@ -1,31 +1,19 @@
-import { isLeft } from "fp-ts/lib/Either";
-import * as t from "io-ts";
-import * as reporter from "io-ts-reporters";
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { getDataSource } from "@app/lib/api/data_sources";
 import { Authenticator, getAPIKey } from "@app/lib/auth";
-import {
-  CoreAPI,
-  CoreAPIDatabaseRow,
-  CoreAPIDatabaseSchema,
-} from "@app/lib/core_api";
+import { CoreAPI, CoreAPIDatabaseTable } from "@app/lib/core_api";
 import { isDevelopmentOrDustWorkspace } from "@app/lib/development";
 import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
-const GetDatabaseSchemaReqBodySchema = t.type({
-  query: t.string,
-});
-
-type QueryDatabaseSchemaResponseBody = {
-  schema: CoreAPIDatabaseSchema;
-  rows: CoreAPIDatabaseRow[];
+type GetDatabaseTableResponseBody = {
+  table: CoreAPIDatabaseTable;
 };
 
 async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<QueryDatabaseSchemaResponseBody>
+  res: NextApiResponse<GetDatabaseTableResponseBody>
 ): Promise<void> {
   const keyRes = await getAPIKey(req);
   if (keyRes.isErr()) {
@@ -65,61 +53,57 @@ async function handler(
     });
   }
 
-  const databaseId = req.query.id;
+  const databaseId = req.query.dId;
   if (!databaseId || typeof databaseId !== "string") {
     return apiError(req, res, {
       status_code: 400,
       api_error: {
         type: "invalid_request_error",
-        message: "Invalid request query: id is required.",
+        message: "The database id is missing.",
+      },
+    });
+  }
+
+  const tableId = req.query.tId;
+  if (!tableId || typeof tableId !== "string") {
+    return apiError(req, res, {
+      status_code: 400,
+      api_error: {
+        type: "invalid_request_error",
+        message: "The table id is missing.",
       },
     });
   }
 
   switch (req.method) {
-    case "POST":
-      const bodyValidation = GetDatabaseSchemaReqBodySchema.decode(req.body);
-      if (isLeft(bodyValidation)) {
-        const pathError = reporter.formatValidationErrors(bodyValidation.left);
-        return apiError(req, res, {
-          status_code: 400,
-          api_error: {
-            type: "invalid_request_error",
-            message: `Invalid request body: ${pathError}`,
-          },
-        });
-      }
-
-      const { query } = bodyValidation.right;
-
-      const queryRes = await CoreAPI.queryDatabase({
+    case "GET":
+      const tableRes = await CoreAPI.getDatabaseTable({
         projectId: dataSource.dustAPIProjectId,
         dataSourceName: dataSource.name,
         databaseId,
-        query,
+        tableId,
       });
-      if (queryRes.isErr()) {
+      if (tableRes.isErr()) {
         logger.error(
           {
-            dataSourceName: dataSource.name,
+            dataSourcename: dataSource.name,
             workspaceId: owner.id,
-            databaseId,
-            error: queryRes.error,
+            error: tableRes.error,
           },
-          "Failed to query database."
+          "Failed to get database table."
         );
         return apiError(req, res, {
           status_code: 500,
           api_error: {
             type: "internal_server_error",
-            message: "Failed to query database.",
+            message: "Failed to get database.",
           },
         });
       }
 
-      const { schema, rows } = queryRes.value;
+      const { table } = tableRes.value;
 
-      return res.status(200).json({ schema, rows });
+      return res.status(200).json({ table });
 
     default:
       return apiError(req, res, {
