+++ front/pages/api/v1/w/[wId]/data_sources/[name]/databases/[dId]/tables/rows/index.ts
@@ -78,7 +78,7 @@ async function handler(
     });
   }
 
-  const databaseId = req.query.id;
+  const databaseId = req.query.dId;
   if (!databaseId || typeof databaseId !== "string") {
     return apiError(req, res, {
       status_code: 400,
@@ -89,7 +89,7 @@ async function handler(
     });
   }
 
-  const tableId = req.query.tableId;
+  const tableId = req.query.tId;
   if (!tableId || typeof tableId !== "string") {
     return apiError(req, res, {
       status_code: 400,
